+++ workspaces/analytics/plugins/analytics-module-matomo/README.md
@@ -21,10 +21,10 @@ If your App is built using the New Frontend System, it's enough to supply this
 module to `createApp()`:
 
 ```tsx
-import matomoModule from '@backstage-community/plugin-analytics-module-matomo/alpha';
+import analyticsProviderMatomoPlugin from '@backstage-community/plugin-analytics-module-matomo/alpha';
 
 const app = createApp({
-  features: [matomoModule],
+  features: [analyticsProviderMatomoPlugin],
 });
 ```
