+++ workspaces/analytics/plugins/analytics-module-matomo/report-alpha.api.md
@@ -3,11 +3,35 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { FrontendModule } from '@backstage/frontend-plugin-api';
+import { AnalyticsImplementationFactory } from '@backstage/frontend-plugin-api';
+import { ExtensionBlueprintParams } from '@backstage/frontend-plugin-api';
+import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
+import { OverridableExtensionDefinition } from '@backstage/frontend-plugin-api';
+import { OverridableFrontendPlugin } from '@backstage/frontend-plugin-api';
 
 // @public (undocumented)
-const matomoModule: FrontendModule;
-export default matomoModule;
+const analyticsProviderMatomoPlugin: OverridableFrontendPlugin<
+  {},
+  {},
+  {
+    'analytics:analytics-module-matomo': OverridableExtensionDefinition<{
+      kind: 'analytics';
+      name: undefined;
+      config: {};
+      configInput: {};
+      output: ExtensionDataRef<
+        AnalyticsImplementationFactory<{}>,
+        'core.analytics.factory',
+        {}
+      >;
+      inputs: {};
+      params: <TDeps extends { [name in string]: unknown }>(
+        params: AnalyticsImplementationFactory<TDeps>,
+      ) => ExtensionBlueprintParams<AnalyticsImplementationFactory<{}>>;
+    }>;
+  }
+>;
+export default analyticsProviderMatomoPlugin;
 
 // (No @packageDocumentation comment for this package)
 ```
