+++ workspaces/analytics/plugins/analytics-module-matomo/src/alpha.ts
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { matomoModule as default } from './module';
+export { analyticsProviderMatomoPlugin as default } from './plugin';
